Removing Resthopper Pipeline
