+++ src/compute.geometry/IO/Utilities.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Net.Http;
+using System.IO;
+using System.Net.Http.Headers;
+using Newtonsoft.Json;
+
+namespace Resthopper.IO
+{
+
+    public static class Utils
+    {
+        public static string Base64Encode(string plainText) {
+            var plainTextBytes = System.Text.Encoding.UTF8.GetBytes(plainText);
+            return System.Convert.ToBase64String(plainTextBytes);
+        }
+    }
+}
